Refactor move uclassify out of text utils
