+++ service/Tests/similarity/TextUtilsTests.py
@@ -1,8 +1,8 @@
 import unittest
 from collections import Counter
 
-from similarity.TextUtils import jaccard_counter_similarity, singleword_similarity, uclassify_topics, \
-    uclassify_similarity, desc_overlap_url
+from similarity.TextUtils import jaccard_counter_similarity, singleword_similarity, desc_overlap_url
+from similarity.UclassifyUtils import uclassify_topics, uclassify_similarity
 
 
 class TextUtilsTests(unittest.TestCase):
@@ -856,17 +856,6 @@ class TextUtilsTests(unittest.TestCase):
         self.assertTrue(0 <= singleword_similarity(self.pinterest_profile, self.instagram_profile) <= 1)
         self.assertTrue(0 <= singleword_similarity(self.twitter_profile, self.instagram_profile) <= 1)
 
-    def test_uclassify_topics(self):
-        text = "Hellö Wörld@Python/?23$%67asdfjkl;[]}{\\';gasdfhjkl"
-        result = uclassify_topics(text)
-        self.assertTrue(len(result.keys()) > 0)
-
-    def test_uclassify_similarity(self):
-        text1 = 'Welcome! Are you completely new to programming? If not then we presume you will be looking for information about why and how to get started with Python.'
-        text2 = 'Before getting started, you may want to find out which IDEs and text editors are tailored to make Python editing easy, browse the list of introductory books, or look at code samples that you might find helpful.'
-        score = uclassify_similarity(text1, text2)
-        self.assertTrue(0 <= score <= 1)
-
     def test_overlapping_urls(self):
         info1 = {'platform': 'instagram', 'username': 'biscuit_boom',
                  'desc': 'Follow me & be rewarded.Fin.;;www.twitter.com/biscuit_boom'}
@@ -878,3 +867,7 @@ class TextUtilsTests(unittest.TestCase):
         info2 = {'platform': 'twitter', 'username': 'Biscuit_Boom',
                  'desc': 'We are a quirky and fun event planning compat'}
         self.assertEqual(0, desc_overlap_url(info1, info2))
+
+
+if __name__ == '__main__':
+    unittest.main()
